Testing geolocation and finalizing write_review.php

+++ upload_review.php
@@ -49,8 +49,8 @@ if (isset($_SESSION['logged_in']) && ($_SESSION['logged_in'])) {
 <?php
 // POST the form inputs into variables to inserted
 $location_name  = filter_input(INPUT_POST, 'location_name');
-$location_id  = filter_input(INPUT_POST, 'location_id');
-$reviewer_id    = filter_input(INPUT_POST, 'reviewer_id');
+$location_id  = (int) filter_input(INPUT_POST, 'location_id');
+$reviewer_id    = (int) filter_input(INPUT_POST, 'reviewer_id');
 $review_title   = filter_input(INPUT_POST, 'reviewTitle');
 $rating        = filter_input(INPUT_POST, 'rating');
 $review_details = filter_input(INPUT_POST, "reviewDetails");
